requirements.txt
aiohttp>=3.8.1
attrs>=21.4.0